refactor: alarmRepository id 데이터타입 수정

+++ src/main/java/com/wanted/preonboarding/ticket/infrastructure/repository/AlarmRepository.java
@@ -3,5 +3,5 @@ package com.wanted.preonboarding.ticket.infrastructure.repository;
 import com.wanted.preonboarding.ticket.domain.entity.Alarm;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlarmRepository extends JpaRepository<Alarm, Long> {
+public interface AlarmRepository extends JpaRepository<Alarm, Integer> {
 }
